Iron out wrinkles; ready for first deployment test!

+++ nitools/qc.py
@@ -1,12 +1,14 @@
 import os
 import os.path as op
 from glob import glob
+import shutil
+
 
 default_args = {
     '--n_procs': 10,
     '--verbose-reports': False,
     '--no-sub': True,
-    '--verbose': False,
+    '--verbose': True,
     '--ica': False,
     '--hmc-afni': True,
     '--hmc-fsl': False,
@@ -20,7 +22,7 @@ default_args = {
 }
 
 
-def run_qc(bids_dir, **mriqc_options):
+def run_qc(bids_dir, export_dir=None, **mriqc_options):
     """ Runs data from BIDS-directory through the MRIQC pipeline.
 
     Parameters
@@ -46,16 +48,25 @@ def run_qc(bids_dir, **mriqc_options):
     default_args.update(mriqc_options)
     mriqc_options = {key: value for key, value in default_args.items() if value}
     options_str = [key + ' ' + str(value) for key, value in mriqc_options.items()]
-    cmd = f'docker run -it --rm -v {bids_dir}:/data:ro -v {out_dir}:/out poldracklab/mriqc:latest /data /out participant ' + ' '.join(options_str).replace(' True', '')
-    cmd += ' --participant_label %s' % ' '.join(participant_labels)
-    #print(cmd)
-    os.system(cmd)
 
+    if participant_labels:
+        cmd = f'docker run -it --rm -v {bids_dir}:/data:ro -v {out_dir}:/out poldracklab/mriqc:latest /data /out participant ' + ' '.join(options_str).replace(' True', '')
+        cmd += ' --participant_label %s' % ' '.join(participant_labels)
+        os.system(cmd)
+    
+    if export_dir is not None:
+        copy_dir = op.join(export_dir, 'qc')
+        if not op.isdir(copy_dir):
+            os.makedirs(copy_dir)
+        
+        proc_sub_data = sorted(glob(op.join(qc_dir, 'sub-*')))
+        done_sub_data = [op.basename(f) for f in sorted(glob(op.join(copy_dir, 'sub-*')))]
+        
+        for f in proc_sub_data:
+            if op.basename(f).split('_')[0] not in done_sub_data:
+                sub_qc_dir = op.join(copy_dir, op.basename(f).split('_')[0])
+                if not op.isdir(sub_qc_dir):
+                    os.makedirs(sub_qc_dir)
 
-if __name__ == '__main__':
-    import yaml
-    cp_file = op.join(op.dirname(__file__), 'data', 'CURRENT_PROJECTS.yml')
-    with open(cp_file, 'r') as cpf:
-        curr_projects = yaml.load(cpf)
+                shutil.copyfile(f, op.join(sub_qc_dir, op.basename(f)))
 
-    run_qc('/media/lukas/data/spinoza-rec/spinoza_testdata/SpinozaTest/bids', **curr_projects['SpinozaTest']['mriqc_options'])
